Update controller and link naming to pascal case
This is because when hosting, the server (linux) cannot read the page
name because controller name isn't the same as link (base_url or site_
url). The Controller is Camel Case with lowercase in second word while
link is all lowered case (snake case).

To solve it, naming must be Pascal Case for both controller and link.